Создание одного RestController с маппингом /management
