+++ src/main/java/com/skypro/sprint1/controller/ManagementController.java
@@ -1,6 +1,7 @@
 package com.skypro.sprint1.controller;
 
 import com.skypro.sprint1.pojo.ServiceInfo;
+import com.skypro.sprint1.service.UserRecommendationService;
 import org.springframework.boot.info.BuildProperties;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -9,12 +10,20 @@ import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequestMapping("/management")
-public class ServiceInfoController {
+public class ManagementController {
 
+    private final UserRecommendationService userRecommendationService;
     private final BuildProperties buildInfo;
 
-    public ServiceInfoController(BuildProperties buildInfo) {
+    public ManagementController(UserRecommendationService userRecommendationService, BuildProperties buildInfo) {
+        this.userRecommendationService = userRecommendationService;
         this.buildInfo = buildInfo;
+    }
+
+    @GetMapping("/clear-caches")
+    public ResponseEntity<String> clearRecommendationCache() {
+        userRecommendationService.clearRecommendationCache();
+        return ResponseEntity.ok("Cache cleared");
     }
 
     @GetMapping("/info")
